+++ backend/accounts/oauth2/models.py
@@ -1,1 +1,3 @@
 from django.db import models
+
+# Create your models here.
