+++ backend/accounts/oauth2/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import FortyTwoLogin
+
+urlpatterns = [
+    path('fortytwo/login/callback/', FortyTwoLogin.as_view(), name='42-login-callback'),
+]
