+++ backend/accounts/two_factor_auth/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TwoFactorAuthConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'accounts.two_factor_auth'
