+++ backend/accounts/two_factor_auth/models.py
@@ -1,1 +1,3 @@
 from django.db import models
+
+# Create your models here.
