+++ backend/accounts/urls.py
@@ -1,13 +1,10 @@
 from django.urls import path, include
 from accounts.views import (
     HelloView,
-    mfaStatus,
-    verifyMFAview,
     CustomLoginView,
-    FortyTwoLogin,
 )
-from .views import get_user_otp_qrcode, CustomTokenVerifyView
-
+from .two_factor_auth.views import get_user_otp_qrcode, mfaStatus, verifyMFAview
+from .views import CustomTokenVerifyView
 
 urlpatterns = [
     path('', include('dj_rest_auth.urls')),
@@ -18,5 +15,5 @@ urlpatterns = [
     path('two-factor-auth/', mfaStatus.as_view(), name='2fa_enable'),
     path('two-factor-auth/verifications/', verifyMFAview, name='2fa_verify'),
     path('custom_login/', CustomLoginView.as_view(), name='login'),
-    path('oauth2/fortytwo/login/callback/', FortyTwoLogin.as_view(), name='callback'),
+    path('oauth2/', include('accounts.oauth2.urls')),
 ]
